+++ src/main/java/kitchenpos/menus/application/dto/request/MenuGroupCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.ui.dto.request;
+package kitchenpos.menus.application.dto.request;
 
 public class MenuGroupCreateRequest {
     private String Name;
