+++ src/main/java/kitchenpos/products/application/ProductService.java
@@ -3,14 +3,14 @@ package kitchenpos.products.application;
 import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuProduct;
 import kitchenpos.menus.domain.MenuRepository;
+import kitchenpos.products.application.dto.request.ProductChangePriceRequest;
+import kitchenpos.products.application.dto.request.ProductCreateRequest;
+import kitchenpos.products.application.dto.response.ProductChangePriceResponse;
+import kitchenpos.products.application.dto.response.ProductCreateResponse;
+import kitchenpos.products.application.dto.response.ProductResponse;
 import kitchenpos.products.domain.Product;
 import kitchenpos.products.domain.ProductRepository;
 import kitchenpos.products.domain.PurgomalumClient;
-import kitchenpos.products.ui.dto.request.ProductChangePriceRequest;
-import kitchenpos.products.ui.dto.request.ProductCreateRequest;
-import kitchenpos.products.ui.dto.response.ProductChangePriceResponse;
-import kitchenpos.products.ui.dto.response.ProductCreateResponse;
-import kitchenpos.products.ui.dto.response.ProductResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
