+++ src/test/java/kitchenpos/menus/application/MenuGroupServiceTest.java
@@ -1,9 +1,9 @@
 package kitchenpos.menus.application;
 
+import kitchenpos.menus.application.dto.request.MenuGroupCreateRequest;
+import kitchenpos.menus.application.dto.response.MenuGroupCreateResponse;
+import kitchenpos.menus.application.dto.response.MenuGroupResponse;
 import kitchenpos.menus.domain.MenuGroupRepository;
-import kitchenpos.menus.ui.dto.request.MenuGroupCreateRequest;
-import kitchenpos.menus.ui.dto.response.MenuGroupCreateResponse;
-import kitchenpos.menus.ui.dto.response.MenuGroupResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
